requirements.txt
attrs==22.2.0
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.0
fonttools==4.38.0
graphviz==0.16
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.6.3
networkx==2.7.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
py-cpuinfo==9.0.0
pydot==1.4.2
pyparsing==3.0.9
pytest==7.2.1
pytest-benchmark==4.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scipy==1.10.0
six==1.16.0
sparse==0.13.0
tomli==2.0.1
tqdm==4.64.1
